+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Bean Pods now replace the final Bean Stalk block instead of being placed above it
 
+### Fixed
+
+- Plate models don't use the Pressure Plate model as a base any more, so it isn't cutting off the outermost 2 pixels
+
 # 1.10.5
 
 ### Fixed